Improved unit tests for open_maybe_zipped function. (#14114)
Implemented "real" tests for open_maybe_zipped that test file content
reading without relying on mocks.
