+++ tests/utils/test_dag_processing.py
@@ -42,7 +42,6 @@ from airflow.utils.dag_processing import (
     DagParsingSignal,
     DagParsingStat,
 )
-from airflow.utils.file import correct_maybe_zipped, open_maybe_zipped
 from airflow.utils.session import create_session
 from airflow.utils.state import State
 from tests.core.test_logging_config import SETTINGS_FILE_VALID, settings_context
@@ -537,65 +536,3 @@ class TestDagFileProcessorAgent(unittest.TestCase):
         processor_agent._process.join()
 
         assert os.path.isfile(log_file_loc)
-
-
-class TestCorrectMaybeZipped(unittest.TestCase):
-    @mock.patch("zipfile.is_zipfile")
-    def test_correct_maybe_zipped_normal_file(self, mocked_is_zipfile):
-        path = '/path/to/some/file.txt'
-        mocked_is_zipfile.return_value = False
-
-        dag_folder = correct_maybe_zipped(path)
-
-        assert dag_folder == path
-
-    @mock.patch("zipfile.is_zipfile")
-    def test_correct_maybe_zipped_normal_file_with_zip_in_name(self, mocked_is_zipfile):
-        path = '/path/to/fakearchive.zip.other/file.txt'
-        mocked_is_zipfile.return_value = False
-
-        dag_folder = correct_maybe_zipped(path)
-
-        assert dag_folder == path
-
-    @mock.patch("zipfile.is_zipfile")
-    def test_correct_maybe_zipped_archive(self, mocked_is_zipfile):
-        path = '/path/to/archive.zip/deep/path/to/file.txt'
-        mocked_is_zipfile.return_value = True
-
-        dag_folder = correct_maybe_zipped(path)
-
-        assert mocked_is_zipfile.call_count == 1
-        (args, kwargs) = mocked_is_zipfile.call_args_list[0]
-        assert '/path/to/archive.zip' == args[0]
-
-        assert dag_folder == '/path/to/archive.zip'
-
-
-class TestOpenMaybeZipped(unittest.TestCase):
-    def test_open_maybe_zipped_normal_file(self):
-        with mock.patch('builtins.open', mock.mock_open(read_data="data")) as mock_file:
-            open_maybe_zipped('/path/to/some/file.txt')
-            mock_file.assert_called_once_with('/path/to/some/file.txt', mode='r')
-
-    def test_open_maybe_zipped_normal_file_with_zip_in_name(self):
-        path = '/path/to/fakearchive.zip.other/file.txt'
-        with mock.patch('builtins.open', mock.mock_open(read_data="data")) as mock_file:
-            open_maybe_zipped(path)
-            mock_file.assert_called_once_with(path, mode='r')
-
-    @mock.patch("zipfile.is_zipfile")
-    @mock.patch("zipfile.ZipFile")
-    @mock.patch("io.TextIOWrapper")
-    def test_open_maybe_zipped_archive(self, mocked_text_io_wrapper, mocked_zip_file, mocked_is_zipfile):
-        mocked_is_zipfile.return_value = True
-        open_return_value = mock.mock_open(read_data="data")
-        instance = mocked_zip_file.return_value
-        instance.open.return_value = open_return_value
-
-        open_maybe_zipped('/path/to/archive.zip/deep/path/to/file.txt')
-
-        mocked_text_io_wrapper.assert_called_once_with(open_return_value)
-        mocked_is_zipfile.assert_called_once_with('/path/to/archive.zip')
-        mocked_zip_file.assert_called_once_with('/path/to/archive.zip', mode='r')
-        instance.open.assert_called_once_with('deep/path/to/file.txt')
